+++ test/shared_map_isolate_test.dart
@@ -458,20 +458,20 @@ void main() {
 
       {
         var sharedMapField2 = SharedMapField(m1.id, sharedStore: store1);
-        expect(identical(sharedMapField2.sharedMapSync, m1), isTrue);
+        expect(identical(sharedMapField2.sharedMap, m1), isTrue);
       }
 
       {
-        expect(sharedMapField.trySharedMapSync?.id, equals(m1.id));
+        expect(sharedMapField.sharedMap.id, equals(m1.id));
 
-        var m2 = sharedMapField.sharedMapSync;
+        var m2 = sharedMapField.sharedMap;
         expect(m2.id, equals(m1.id));
 
         expect(identical(m2, m1), isTrue);
       }
 
       var va5 = await Isolate.run<int?>(() async {
-        var m3 = await sharedMapField.sharedMap;
+        var m3 = sharedMapField.sharedMap;
 
         var va5 = await m3.get('a');
         if (va5 != 111) throw StateError("Expected: 111 ; got: $va5");